refactor: Seperate car count validation and add test case
