+++ src/domain/validate.js
@@ -1,9 +1,13 @@
 import { ERROR_MESSAGE } from "../constants.js";
 
+export const validateCarCount = (carNames) => {
+  if (carNames.length < 2) throw Error(ERROR_MESSAGE.TOO_FEW_CARS);
+  if (carNames.length > 10) throw Error(ERROR_MESSAGE.TOO_MANY_CARS);
+}
+
 export const validateCarNames = (carNames) => {
   if (carNames.some(name => name.length == 0)) throw new Error(ERROR_MESSAGE.CAR_NAME_TOO_SHORT);
   if (carNames.some(name => name.length > 5)) throw new Error(ERROR_MESSAGE.CAR_NAME_TOO_LONG);
-  if (carNames.length > 10) throw Error(ERROR_MESSAGE.TOO_MANY_CARS);
   if (carNames.length !== new Set(carNames).size) throw Error(ERROR_MESSAGE.CAR_NAME_DUPLICATE);
 };
